moved Grid to GameObjects and added some comments
